Comment model changes
